requirements.txt
Flask==3.0.3
python-escpos==3.1
Pillow==10.3.0
